BoatDamageを分離
