+++ Assets/Scripts/BoatDamage.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: efb7577fd748648efb6cd64524678686
+timeCreated: 1494116356
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
